locuras con deivid

+++ DLP_1/src/ast/expressions/RealLiteral.java
@@ -4,9 +4,9 @@ import ast.ConcreteASTNode;
 
 public class RealLiteral extends ConcreteASTNode implements Expression {
 
-	public long real;
+	public double real;
 
-	public RealLiteral(int row, int column, long l) {
+	public RealLiteral(int row, int column, double l) {
 		super(row, column);
 		real = l;
 	}
